requirements.txt
django
djangorestframework
social-auth-app-django
drf-yasg
Pillow
